+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - The deprecated `go.opentelemetry.io/otel/metric/instrument` package is removed.
   Use `go.opentelemetry.io/otel/metric` instead. (#4055)
 
+### Fixed
+
+- Fix build for BSD based systems in `go.opentelemetry.io/otel/sdk/resource`. (#4077)
+
 ## [1.16.0-rc.1/0.39.0-rc.1] 2023-05-03
 
 This is a release candidate for the v1.16.0/v0.39.0 release.
@@ -115,6 +119,9 @@ See our [versioning policy](VERSIONING.md) for more information about these stab
 - Configuration for each metric instrument in `go.opentelemetry.io/otel/sdk/metric/instrument`. (#3895)
 - The internal logging introduces a warning level verbosity equal to `V(1)`. (#3900)
 - Added a log message warning about usage of `SimpleSpanProcessor` in production environments. (#3854)
+- The `go.opentelemetry.io/otel/semconv/v1.19.0` package. The package contains
+semantic conventions from the `v1.19.0` version of the OpenTelemetry
+specification. (#3846)
 
 ### Changed